Move butter_bandpass into utils.utils.
