+++ xcp_d/interfaces/filtering.py
@@ -5,7 +5,6 @@
 .. testsetup::
 # will comeback
 """
-import numpy as np
 from nipype import logging
 from nipype.interfaces.base import (
     BaseInterfaceInputSpec,
@@ -14,9 +13,9 @@ from nipype.interfaces.base import (
     TraitedSpec,
     traits,
 )
-from scipy.signal import butter, filtfilt
 
 from xcp_d.utils.filemanip import fname_presuffix
+from xcp_d.utils.utils import butter_bandpass
 from xcp_d.utils.write_save import read_ndata, write_ndata
 
 LOGGER = logging.getLogger('nipype.interface')
@@ -106,42 +105,3 @@ class FilteringData(SimpleInterface):
             filename=self._results['filtered_file'],
             mask=self.inputs.mask)
         return runtime
-
-
-def butter_bandpass(data, fs, lowpass, highpass, order=2):
-    """Apply a Butterworth bandpass filter to data.
-
-    Parameters
-    ----------
-    data : numpy.ndarray
-        Voxels/vertices by timepoints dimension.
-    fs : float
-        Sampling frequency. 1/TR(s).
-    lowpass : float
-        frequency
-    highpass : float
-        frequency
-    order : int
-        The order of the filter. This will be divided by 2 when calling scipy.signal.butter.
-
-    Returns
-    -------
-    filtered_data : numpy.ndarray
-        The filtered data.
-    """
-    nyq = 0.5 * fs  # nyquist frequency
-
-    # normalize the cutoffs
-    lowcut = np.float(highpass) / nyq
-    highcut = np.float(lowpass) / nyq
-
-    b, a = butter(order / 2, [lowcut, highcut], btype='band')  # get filter coeff
-
-    filtered_data = np.zeros(data.shape)  # create something to populate filtered values with
-
-    # apply the filter, loop through columns of regressors
-    for ii in range(filtered_data.shape[0]):
-        filtered_data[ii, :] = filtfilt(b, a, data[ii, :], padtype='odd',
-                                        padlen=3 * (max(len(b), len(a)) - 1))
-
-    return filtered_data
